Renamed pick.js to picker.js to match class name Picker (#2726)
* Renamed pick.js to picker.js to match class name Picker

* remove unused library
