+++ src/deprecated.js
@@ -62,7 +62,7 @@ import { Mesh } from './scene/mesh.js';
 import { MeshInstance, Command } from './scene/mesh-instance.js';
 import { Model } from './scene/model.js';
 import { ParticleEmitter } from './scene/particle-system/particle-emitter.js';
-import { Picker } from './scene/pick.js';
+import { Picker } from './scene/picker.js';
 import { Scene } from './scene/scene.js';
 import { Skin } from './scene/skin.js';
 import { SkinInstance } from './scene/skin-instance.js';
